+++ HISTORY.md
@@ -10,6 +10,7 @@
 - Added support for VHDL-2008 "all" sensitised processes.
 - Added support for ports and generics in block statements.
 - Added support for the 'BASE attribute.
+- Type name now allowed in element association choice (#407).
 
 ## 1.5 - 2020-07-19
 - IEEE library sources are now distributed